fix build after merge
