+++ core/src/mast/node/mod.rs
@@ -1,12 +1,11 @@
 mod basic_block_node;
-use alloc::{boxed::Box, string::ToString, vec::Vec};
+use alloc::{boxed::Box, vec::Vec};
 use core::fmt;
 
 pub use basic_block_node::{
     BasicBlockNode, OpBatch, OperationOrDecorator, BATCH_SIZE as OP_BATCH_SIZE,
     GROUP_SIZE as OP_GROUP_SIZE,
 };
-use num_traits::ToBytes;
 
 mod call_node;
 pub use call_node::CallNode;
@@ -21,13 +20,7 @@ mod join_node;
 pub use join_node::JoinNode;
 
 mod split_node;
-use miden_crypto::{
-    hash::{
-        blake::{Blake3Digest, Blake3_256},
-        rpo::RpoDigest,
-    },
-    Felt,
-};
+use miden_crypto::{hash::rpo::RpoDigest, Felt};
 use miden_formatting::prettier::{Document, PrettyPrint};
 pub use split_node::SplitNode;
 
@@ -37,7 +30,7 @@ pub use loop_node::LoopNode;
 use super::{DecoratorId, MastForestError};
 use crate::{
     mast::{MastForest, MastNodeId},
-    Decorator, DecoratorList, Operation,
+    DecoratorList, Operation,
 };
 
 // MAST NODE
@@ -205,49 +198,6 @@ impl MastNode {
         }
     }
 
-    /// Returns the Blake3 hash of this node, to be used for equality testing.
-    ///
-    /// Specifically, two nodes with the same MAST root but different decorators will have a
-    /// different hash.
-    pub fn eq_hash(&self) -> Blake3Digest<32> {
-        match self {
-            MastNode::Block(node) => {
-                let mut bytes_to_hash = node.digest().as_bytes().to_vec();
-
-                for (idx, decorator) in node.decorators() {
-                    bytes_to_hash.extend(idx.to_le_bytes());
-
-                    match decorator {
-                        Decorator::Advice(advice) => {
-                            bytes_to_hash.extend(advice.to_string().as_bytes())
-                        },
-                        Decorator::AsmOp(asm_op) => {
-                            bytes_to_hash.extend(asm_op.context_name().as_bytes());
-                            bytes_to_hash.extend(asm_op.op().as_bytes());
-                            bytes_to_hash.push(asm_op.num_cycles());
-                        },
-                        Decorator::Debug(debug) => {
-                            bytes_to_hash.extend(debug.to_string().as_bytes())
-                        },
-                        Decorator::Event(event) => {
-                            bytes_to_hash.extend(event.to_le_bytes());
-                        },
-                        Decorator::Trace(trace) => {
-                            bytes_to_hash.extend(trace.to_le_bytes());
-                        },
-                    }
-                }
-
-                Blake3_256::hash(&bytes_to_hash)
-            },
-            MastNode::Join(node) => Blake3_256::hash(&node.digest().as_bytes()),
-            MastNode::Split(node) => Blake3_256::hash(&node.digest().as_bytes()),
-            MastNode::Loop(node) => Blake3_256::hash(&node.digest().as_bytes()),
-            MastNode::Call(node) => Blake3_256::hash(&node.digest().as_bytes()),
-            MastNode::Dyn(node) => Blake3_256::hash(&node.digest().as_bytes()),
-            MastNode::External(node) => Blake3_256::hash(&node.digest().as_bytes()),
-        }
-    }
     /// Returns the decorators to be executed before this node is executed.
     pub fn before_enter(&self) -> &[DecoratorId] {
         use MastNode::*;
